Error de persistencia resolvido

+++ src/main/java/com/moisesfelix/workshopmongodb1/config/Instantiation.java
@@ -3,6 +3,7 @@ package com.moisesfelix.workshopmongodb1.config;
 import java.util.Arrays;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Configuration;
 
@@ -12,18 +13,23 @@ import com.moisesfelix.workshopmongodb1.repositories.UserRepository;
 @Configuration
 public class Instantiation implements CommandLineRunner{
 	
+	@Value("${user}")
+	private String ambiente;
 	@Autowired
 	private UserRepository userRepository;
 
 	@Override
 	public void run(String... args) throws Exception {
-		userRepository.deleteAll();
-		
+		System.out.println(ambiente.contains("test"));
+		System.out.println(ambiente);
+		if(ambiente.contains("test")) {
+			userRepository.deleteAll();
 		User maria = new User(null, "Maria Brown", "maria@gmail.com"); 
 		User alex = new User(null, "Alex Green", "alex@gmail.com"); 
 		User bob = new User(null, "Bob Grey", "bob@gmail.com");
 		
 		userRepository.saveAll(Arrays.asList(maria, alex, bob));
+		}
 	}
 
 }
